test: :white_check_mark: add intergration test demo
