+++ services/user/test/index.js
@@ -1,0 +1,6 @@
+/* eslint-disable import/no-extraneous-dependencies */
+const test = require('ava');
+const debug = require('debug');
+
+exports.test = test;
+exports.debug = debug('gql:service:user');
